requirements.txt
Django==3.2
mysqlclient==2.0.3
python-dotenv==1.0.0